requirements.txt
django
django_hint
pillow
django-ckeditor
htmlmin
beautifulsoup4